requirements.txt
absl-py==0.1.10
astor==0.6.2
backports-abc==0.5
bleach==1.5.0
cycler==0.10.0
Cython==0.28.1
decorator==4.2.1
entrypoints==0.2.3
flake8==3.5.0
gast==0.2.0
grpcio==1.10.0
html5lib==0.9999999
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.2.0
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.1
numpy==1.14.0
opencv-python==3.4.0.12
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.3
pyzmq==17.0.0
qtconsole==4.3.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.7.0
tensorflow-gpu==1.7.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.1
traitlets==4.3.2
typing==3.6.4
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.2.0
